frontend/openapi/src/main/java/org/dotwebstack/framework/frontend/openapi/entity/schema/FloatSchemaMapper.java
package org.dotwebstack.framework.frontend.openapi.entity.schema;

import com.google.common.collect.ImmutableSet;
import io.swagger.v3.oas.models.media.NumberSchema;
import io.swagger.v3.oas.models.media.Schema;
import java.util.Set;
import lombok.NonNull;
import org.dotwebstack.framework.frontend.openapi.OpenApiSpecificationExtensions;
import org.eclipse.rdf4j.model.IRI;
import org.eclipse.rdf4j.model.Literal;
import org.eclipse.rdf4j.model.vocabulary.XMLSchema;
import org.springframework.stereotype.Service;

@Service
class FloatSchemaMapper extends AbstractSchemaMapper<NumberSchema, Float> {

  private static final Set<IRI> SUPPORTED_TYPES = ImmutableSet.of(XMLSchema.FLOAT);
  private static final Set<String> SUPPORTED_VENDOR_EXTENSIONS = ImmutableSet.of(
      OpenApiSpecificationExtensions.LDPATH, OpenApiSpecificationExtensions.CONSTANT_VALUE);

  @Override
  protected Set<String> getSupportedVendorExtensions() {
    return SUPPORTED_VENDOR_EXTENSIONS;
  }

  @Override
  public boolean supports(@NonNull Schema schema) {
    return schema instanceof NumberSchema && schema.getFormat().equals("float");
  }

  @Override
  protected Float convertLiteralToType(Literal literal) {
    return literal.floatValue();
  }

  @Override
  protected Set<IRI> getSupportedDataTypes() {
    return SUPPORTED_TYPES;
  }
}
